Объединил логику RPN Encoder/ Decoder в RPNConverter.
